feat: Augment status endpoint for stoppable action
in order to be able to report to the platform when an action was stopped
by the extension, we augment stoppable actions with a status endpoint,
that will return an error when an execution was stopped by the
extension either via SIGUSR1, SIGTERM.
The extension sdk is monitoring the calls to the status endpoint. If the
calls are missing for 4 * status interval. Then the action if stopped.