fix: prevent extension heartbeat from stopping randomly
This issue was caused by a race condition and, hence, hard to attribute to a
specific thing. In situations in which the heartbeat was constantly missed, the
heartbeat channel was reported as full, so no new messages were added, and no
reads from the channel were performed.

This change adds some jitter to reduce missed heartbeats due to network and
processing time and changes the heartbeat queue to a buffered one so that new
messages can be added.